% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regioneReloaded_data.R
\docType{data}
\name{AlienGenome}
\alias{AlienGenome}
\title{AlienGenome}
\format{
An objects of class \link{GRanges}.
}
\usage{
data(cw_Alien)
}
\description{
The Alien Genome is an artificial genomic coordinates system for the purposes
of testing and demonstrating the functions of regioneReload with a low
computing time.
}
\details{
The Alien Genome consists of four chromosomes and is generated by the
following code:

\preformatted{AlienGenome <-
toGRanges(data.frame(
  chr = c("AlChr1", "AlChr2", "AlChr3", "AlChr4"),
  start = c(rep(1, 4)),
  end = c(2e6, 1e6, 5e5, 1e5)
))}
}
\keyword{datasets}
