% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cw_Alien-data.R
\docType{data}
\name{AlienRSList}
\alias{AlienRSList}
\title{AlienRSList_narrow}
\format{
A list of \link{GRanges} objects.
}
\usage{
data(cw_Alien)
}
\description{
Description
}
\details{
Details
}
\keyword{datasets}
