% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regioneReloaded_data.R
\docType{data}
\name{AlienRSList_narrow}
\alias{AlienRSList_narrow}
\title{AlienRSList_narrow}
\format{
A list of \link[GenomicRanges:GRanges-class]{GRanges} objects.
}
\usage{
data(cw_Alien)
}
\description{
List of region sets (as \link[GenomicRanges:GRanges-class]{GRanges}) on the \link{AlienGenome}.
}
\details{
This region sets are generated for the purpose of demonstrating the functions
of RegioneReloaded with a low computing time and "predictable" associations. The
regions are generated with by combining \code{\link[=createRandomRegions]{createRandomRegions()}} and
\code{\link[=similarRegionSet]{similarRegionSet()}} so that there is a known overlap between certain region
sets. To see a full description of this sample data and the code used to
generate it, see the RegioneReloaded vignette.
}
\keyword{datasets}
