% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleanCrosswiseMatrix.R
\name{cleanCrosswiseMatrix}
\alias{cleanCrosswiseMatrix}
\title{cleanCrosswiseMatrix}
\usage{
cleanCrosswiseMatrix(GM, GM_pv, pvcut, scale, subEX)
}
\arguments{
\item{GM}{matrix,  numerical matrix of z-scores.}

\item{GM_pv}{matrix, numerical matrix of pvalues.}

\item{pvcut}{numeric, the z-score value is substituted by \code{subEX} (0 by default) for all the associations with an adj.pvalue (as calculated in \code{\link[=crosswisePermTest]{crosswisePermTest()}}) higher than \code{pvcut}. (default = 0.05)}

\item{scale}{logical, if TRUE the matrix will be scaled. (default = FALSE)}

\item{subEX}{numeric, value used to substitute the z-score values when the associated pvalue is higher than \code{pvcut}. (default = 0)}
}
\value{
a matrix filtered for a matrix of pvalue
}
\description{
Clean and scale a matrix from a genoMatriXeR object
}
\seealso{
\code{\link[=makeCrosswiseMatrix]{makeCrosswiseMatrix()}}
}
\keyword{internal}
