% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regioneRld_themes.R
\docType{data}
\name{colvec}
\alias{colvec}
\title{Themes used in the regioneReloaded package}
\format{
An object of class \code{character} of length 5.
}
\usage{
mendel_theme(
  base_size = 11,
  base_family = ""
)
}
\arguments{
\item{base_size}{base font size, given in pts.}

\item{base_family}{base font family}
}
\description{
Ggplot2 themes created for the package.
}
\keyword{function}
\keyword{internal}
