% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createUniverse.R
\name{createUniverse}
\alias{createUniverse}
\title{createUniverse}
\usage{
createUniverse(Alist, joinR = TRUE)
}
\arguments{
\item{Alist}{list of regions set in a format accepted for \link[regioneR:reginoeR]{regioneR::reginoeR}}

\item{joinR}{logical, if TRUE all the regions will be joined using the function \code{\link[regioneR:joinRegions]{regioneR::joinRegions()}}.(default == TRUE)}
}
\value{
A list of \link[=GenomicRanges-class]{GenomicRanges} objects
}
\description{
Create the universe parameter for \code{\link[regioneR:resampleRegions]{regioneR::resampleRegions()}} using all unique regions present in Alist.
}
\examples{

data("cw_Alien")

universe <- createUniverse(AlienRSList_narrow)

}
