% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crosswiseMatrix.R
\name{crosswiseMatrix}
\alias{crosswiseMatrix}
\title{crosswiseMatrix}
\usage{
crosswiseMatrix(mPT, zs.type='norm_zscore',...)
}
\arguments{
\item{mPT}{an object of class \link[=genoMatriXeR-class]{genoMatriXeR}.}

\item{zs.type}{character, z-score type to use to generate the matrix, either raw z-score ("zscore") or normalized z-score ("norm_zscore"). (default = "norm_zscore")}

\item{...}{further arguments to be passed to other methods.}
}
\value{
a numeric matrix
}
\description{
Create a matrix from a genoMatriXeR object.
}
\examples{

data("cw_Alien")

matrix_CW_Alien <- crosswiseMatrix(cw_Alien_ReG)

}
