% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cw_Alien-data.R
\docType{data}
\name{cw_Alien}
\alias{cw_Alien}
\title{Alien Genome crosswise matrix using \code{\link{randomizeRegions}} , \code{\link{circularRandomizeRegions}},
\code{\link{resampleRegionS}}, \code{\link{resampleGenome}}    functions as permutation strategies}
\format{
Four objects of class \code{"genoMatriXeR"}; see \code{\link{makeCrosswiseMatrix}}.
}
\usage{
data(cw_Alien)
}
\description{
CrossWise matrix calculated using 39 different regionset vs itself
(description of Region set can be find here:"")
}
\keyword{datasets}
