% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cw_Alien-data.R
\docType{data}
\name{cw_Alien_RaR}
\alias{cw_Alien_RaR}
\title{cw_Alien_RaR}
\format{
An objects of class \link[=genoMatriXeR-class]{genoMatriXeR}; see \code{\link[=makeCrosswiseMatrix]{makeCrosswiseMatrix()}}.
}
\usage{
data(cw_Alien)
}
\description{
Alien Genome crosswise matrix using \code{\link[regioneR:randomizeRegions]{regioneR::randomizeRegions()}}
function a permutation strategy.
}
\keyword{datasets}
