% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regioneReloaded_data.R
\docType{data}
\name{cw_Alien_ReG}
\alias{cw_Alien_ReG}
\title{cw_Alien_ReG}
\format{
An objects of class \link[=genoMatriXeR-class]{genoMatriXeR}; see \code{\link[=makeCrosswiseMatrix]{makeCrosswiseMatrix()}}.

An objects of class \link[=multiLocalZScore-class]{multiLocalZScore}; see \code{\link[=makeLZMatrix]{makeLZMatrix()}}.
}
\usage{
data(cw_Alien)

data(cw_Alien)
}
\description{
Alien Genome crosswise matrix using \code{\link[regioneR:resampleGenome]{regioneR::resampleGenome()}} function as permutations
trategy. Alist = AlienRSList_narrow, Blist = AlienRSList_narrow

Alien Genome multiLocalZScore calculated for regA regionset from AlienRSList_narrow using \code{\link[regioneR:resampleGenome]{regioneR::resampleGenome()}} function as permutation s
trategy.
}
\keyword{datasets}
