% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regioneReloaded_data.R
\docType{data}
\name{cw_Alien_ReG_no_Square}
\alias{cw_Alien_ReG_no_Square}
\title{cw_Alien_ReG_no_Square}
\format{
An objects of class \link[=genoMatriXeR-class]{genoMatriXeR}; see \code{\link[=makeCrosswiseMatrix]{makeCrosswiseMatrix()}}.
}
\usage{
data(cw_Alien)
}
\description{
Alien Genome crosswise matrix using \code{\link[regioneR:resampleGenome]{regioneR::resampleGenome()}} function as permutations
trategy. Alist = AlienRSList_narrow, Blist = AlienRSList_broad
}
\keyword{datasets}
