% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regioneReloaded_data.R
\docType{data}
\name{cw_Alien_ReR}
\alias{cw_Alien_ReR}
\title{cw_Alien_ReR}
\format{
An objects of class \link[=genoMatriXeR-class]{genoMatriXeR}; see \code{\link[=makeCrosswiseMatrix]{makeCrosswiseMatrix()}}.

An objects of class \link[=multiLocalZScore-class]{multiLocalZScore}; see \code{\link[=makeLZMatrix]{makeLZMatrix()}}.
}
\usage{
data(cw_Alien)

data(cw_Alien)
}
\description{
Alien Genome crosswise matrix using \code{\link[regioneR:resampleRegions]{regioneR::resampleRegions()}} function a permutation
strategy.

Alien Genome multiLocalZScore calculated for regA regionset from AlienRSList_narrow using \code{\link[=resampleGenome]{resampleGenome()}} function as permutation s
trategy.
}
\keyword{datasets}
