% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cw_Real-data.R
\docType{data}
\name{cw_Real}
\alias{cw_Real}
\title{cw_Real}
\format{
Two objects of class \code{"genoMatriXeR"} and \code{"multiLocalZscore"}; see \code{\link{makeCrosswiseMatrix}}.
}
\usage{
data(cw_Real)
}
\description{
HepG2 TF crosswise matrix using \code{\link{resampleGenome}} functions as permutation strategies
}
\details{
CrossWise matrix calculated using 261 different regionset vs itself
(description of Region set can be find here:"")
}
\keyword{datasets}
