% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DFformLZ.R
\name{DFfromLZ}
\alias{DFfromLZ}
\title{Dataframe from MultiLocalZScore object}
\usage{
DFfromLZ( mLZ, RS )
}
\arguments{
\item{mLZ}{object class MultilocalZScore}

\item{RS}{string name of single region set present in the mLZ object}
}
\value{
a data.frame from \linkS4class{multiLocalZScore}
}
\description{
Dataframe from MultiLocalZScore object
}
\keyword{function}
\keyword{internal}
