% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getHClust.R
\name{getHClust}
\alias{getHClust}
\title{getHClust}
\usage{
getHClust( rR, hctype = "rows")
}
\arguments{
\item{rR}{A genoMatriXeR or multiLocalZScore object.}

\item{hctype}{character. Can be "rows" or "cols". (default= "cols")}
}
\value{
an object of class \link{hclust}
}
\description{
get Object of class \link{hclust} from  \link[=genoMatriXeR_class]{genoMatriXeR} or \link[=multiLocalZScore_class]{multiLocalZScore}
}
\examples{

data("cw_Alien")

hc <- getHClust(cw_Alien_ReG)

plot(hc)

}
\seealso{
\link[=genoMatriXeR-class]{genoMatriXeR}, \link[=multiLocalZScore_class]{multiLocalZScore}, \link{hclust}
}
