% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMatrix.R
\name{getMatrix}
\alias{getMatrix}
\title{Get Matrix}
\usage{
getMatrix(rR, vebose = TRUE)
}
\arguments{
\item{rR}{genoMatriXeR or multiLocalZScore object}
}
\value{
a numeric Matrix
}
\description{
Return matrix from an genoMatriXeR or multiLocalZScore object
}
\examples{

data("cw_Alien")

cw_Alien_RaR <- makeCrosswiseMatrix(cw_Alien_RaR)
GM <- getMatrix(cw_Alien_RaR)

}
