% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMultiEvaluation.R
\name{getMultiEvaluation}
\alias{getMultiEvaluation}
\title{getMultiEvaluation}
\usage{
getMultiEvaluation( rR, namesRS = NULL)
}
\arguments{
\item{rR}{A genoMatriXeR or multiLocalZScore object.}

\item{namesRS}{a vector of names. (default = NA)}
}
\value{
If rR is a \link[=genoMatriXeR-class]{genoMatriXeR} object, a list of data frames resuming the associations results.
If rR is a \link[=multiLocalZScore-class]{multiLocalZScore} object, a list of two elements: "resumeTable" that is a data frame
summarizing the associations and "shifts", a list of shifts computed from \code{\link[=multiLocalZscore]{multiLocalZscore()}} function for the elements
indicated in the nameRS vector.
}
\description{
Get \code{multiEvaluation} slot from \link[=genoMatriXeR-class]{genoMatriXeR} or \link[=multiLocalZScore-class]{multiLocalZScore} class.
}
\examples{

data("cw_Alien")

mevs <- getMultiEvaluation(cw_Alien_ReG, names = "regA")

mevs

}
\seealso{
\link[=genoMatriXeR-class]{genoMatriXeR}, \link[=multiLocalZScore-class]{multiLocalZScore}
}
