% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regioneReloaded_methods.R
\name{gmxrMatrix<-}
\alias{gmxrMatrix<-}
\title{Set matrix slot value of a genoMatriXeR object}
\usage{
gmxrMatrix(x) <- value
}
\arguments{
\item{x}{A genoMatriXeR object.}

\item{value}{Value to assign to the matrix slot.}
}
\value{
A genoMatriXeR object.
}
\description{
Set matrix slot value of a genoMatriXeR object
}
\keyword{internal}
