% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regioneReloaded_methods.R
\name{gmxrMatrix}
\alias{gmxrMatrix}
\title{Export matrix slot from genoMatriXeR objects}
\usage{
gmxrMatrix(x)
}
\arguments{
\item{x}{A genoMatriXeR object.}
}
\value{
A printed output for genoMatriXeR matrix slot.
}
\description{
Export matrix slot from genoMatriXeR objects
}
\seealso{
\link[=genoMatriXeR-class]{genoMatriXeR}
}
