% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regioneReloaded_methods.R
\name{gmxrMultiOverlaps}
\alias{gmxrMultiOverlaps}
\title{Export multiOverlaps slot from genoMatriXeR objects}
\usage{
gmxrMultiOverlaps(x)
}
\arguments{
\item{x}{A genoMatriXeR object.}
}
\value{
A printed output for genoMatriXeR multiOverlaps slot.
}
\description{
Export multiOverlaps slot from genoMatriXeR objects
}
\seealso{
\link[=genoMatriXeR-class]{genoMatriXeR}
}
