% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regioneReloaded_methods.R
\name{gmxrParam}
\alias{gmxrParam}
\title{Export parameters from genoMatriXeR objects}
\usage{
gmxrParam(x)
}
\arguments{
\item{x}{A genoMatriXeR object.}
}
\value{
A printed output for genoMatriXeR parameters slot.
}
\description{
Export parameters from genoMatriXeR objects
}
\seealso{
\link[=genoMatriXeR-class]{genoMatriXeR}
}
