% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regioneReloaded_theme.R
\name{mendel_theme}
\alias{mendel_theme}
\title{Themes used in the regioneReloaded package}
\usage{
mendel_theme(
  base_size = 11,
  base_family = ""
)
}
\arguments{
\item{base_size}{base font size, given in pts.}

\item{base_family}{base font family}
}
\description{
Ggplot2 themes created for the package.
}
\keyword{function}
\keyword{internal}
