% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regioneReloaded_data.R
\docType{data}
\name{mLZ_regA_ReG_br}
\alias{mLZ_regA_ReG_br}
\title{mLZ_regA_ReG_br}
\format{
An object of class \link[=multiLocalZScore-class]{multiLocalZScore}
}
\usage{
data(cw_Alien)
}
\description{
Alien Genome multiLocalZScore calculated for regA regionset from AlienRSList_broad using \code{\link[regioneR:resampleGenome]{regioneR::resampleGenome()}} function as permutation s
trategy.
}
