% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regioneReloaded_data.R
\docType{data}
\name{mLZ_regD_ReG}
\alias{mLZ_regD_ReG}
\title{mLZ_regD_ReG}
\format{
An object of class \link[=multiLocalZScore-class]{multiLocalZScore}
}
\usage{
data(cw_Alien)
}
\description{
Placeholder
}
