% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regioneReloaded_methods.R
\name{mlzsMatrix<-}
\alias{mlzsMatrix<-}
\title{Set matrix slot value of a multiLocalZScore object.}
\usage{
mlzsMatrix(x) <- value
}
\arguments{
\item{x}{A multiLocalZScore object.}

\item{value}{Value to assign to the multiLocalZScore slot.}
}
\value{
A multiLocalZScore object.
}
\description{
Set matrix slot value of a multiLocalZScore object.
}
\keyword{internal}
