% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regioneReloaded_methods.R
\name{mlzsMatrix}
\alias{mlzsMatrix}
\title{Export matrix from multiLocalZScore objects}
\usage{
mlzsMatrix(x)
}
\arguments{
\item{x}{A multiLocalZScore object.}
}
\value{
A printed output for multiLocalZScore matrix slot.
}
\description{
Export matrix from multiLocalZScore objects
}
\seealso{
\link[=multiLocalZScore-class]{multiLocalZScore}
}
