% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regioneReloaded_methods.R
\name{mlzsMultiLocalZscores}
\alias{mlzsMultiLocalZscores}
\title{Export multiLocalZscores slot from multiLocalZScore objects}
\usage{
mlzsMultiLocalZscores(x)
}
\arguments{
\item{x}{A multiLocalZScore object.}
}
\value{
A printed output for multiLocalZScore multiLocalZscores slot.
}
\description{
Export multiLocalZscores slot from multiLocalZScore objects
}
\seealso{
\link[=multiLocalZScore-class]{multiLocalZScore}
}
