% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regioneReloaded_methods.R
\name{mlzsParam<-}
\alias{mlzsParam<-}
\title{Set parameters slot value of a multiLocalZScore object}
\usage{
mlzsParam(x) <- value
}
\arguments{
\item{x}{A multiLocalZScore object.}

\item{value}{Value to assign to the parameters slot.}
}
\value{
A multiLocalZScore object.
}
\description{
Set parameters slot value of a multiLocalZScore object
}
\keyword{internal}
