% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regioneReloaded_methods.R
\name{mlzsParam}
\alias{mlzsParam}
\title{Export parameters from multiLocalZScore objects}
\usage{
mlzsParam(x)
}
\arguments{
\item{x}{A multiLocalZScore object.}
}
\value{
parameters slot of a multiLocalZScore object.
}
\description{
Export parameters from multiLocalZScore objects
}
\seealso{
\link[=multiLocalZScore-class]{multiLocalZScore}
}
\keyword{internal}
