% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mLZS_class.R
\docType{class}
\name{multiLocalZScore-class}
\alias{multiLocalZScore-class}
\alias{mLZS}
\title{multiLocalZScore Class}
\description{
An S4 class for "multiLocalZScore" object.
}
\section{Slots}{

\describe{
\item{\code{parameters}}{List of parameters used to create the object}

\item{\code{multiLocalZscores}}{Results of multiple pairwise permutation tests on shifted region sets generated with \code{\link[=multiLocalZscore]{multiLocalZscore()}}.}

\item{\code{matrix}}{List of numerical matrices containing local z-scores and correlation values generated with \code{\link[=makeLZMatrix]{makeLZMatrix()}}.}
}}

