% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiLocalZscore.R
\name{multiLocalZscore}
\alias{multiLocalZscore}
\title{Multiple Local Z-Score test}
\usage{
multiLocalZscore( A, Blist, ranFUN = "randomizeRegions", evFUN = "numOverlaps", sampling = FALSE,
min_sampling = 5000, fraction = 0.15, universe = NULL, window = 1000, step = 100, adj_pv_method = "BH", min_regions = 1000,
max_pv = 0.05, genome = "hg19", ...)
}
\arguments{
\item{A}{Region Set of any accepted formats by  c package
(\code{\link{GenomicRanges}}, \code{\link{data.frame}} etc...)}

\item{Blist}{list of Region Set of any accepted formats by \href{https://bioconductor.org/packages/release/bioc/html/regioneR.html}{regioneR} package
(\code{\link{GenomicRanges}}, \code{\link{data.frame}} etc...)}

\item{ranFUN}{Function, choose the randomization strategy used for the test (default = "randomizeRegions")
for details see \href{https://bioconductor.org/packages/release/bioc/html/regioneR.html}{regioneR}}

\item{evFUN}{Function, choose the evaluation strategy used for the test (default = "numOverlaps")}

\item{sampling}{Boolean, if is true the function will use only a sample of each element of Alist to perform the test.(default = FALSE)}

\item{min_sampling}{Numeric, minimum number of regions in the region set that permit a sampling. (default = 5000)}

\item{fraction}{Numeric, if sampling == TRUE is the fraction of the region sets used to perform the test. (default = 0.15)}

\item{universe}{Region Set of any accepted formats by  \href{https://bioconductor.org/packages/release/bioc/html/regioneR.html}{regioneR}, using only when resamplinRegions function is
selected (default = NULL)}

\item{window}{window (number of base pairs) in which will be estimated the local Z-score. (default = 1000)}

\item{step}{step (number of base pairs) in which will be estimated the local Z-score. (default = 100)}

\item{adj_pv_method}{Charachter, the method used for the calculation of the adjusted p-value,
to choose between the options of \code{\link{p.adjust}}. (default = "BH")}

\item{min_regions}{minimun regions accepted after the sampling. (default = 1000)}

\item{max_pv}{Numeric, the z-scores associate a p-values higher of this parameter will be transform in 0. (default =0.05)}

\item{genome}{Charachter or GenomicRanges, genome used to compute the randomization. (default = "hg19")}

\item{...}{further arguments to be passed to other methods.}
}
\value{
A object of class \link[=multiLocalZScore-class]{multiLocalZScore} containing three slots

\itemize{
\item \bold{\code{@parameters}}
\item \bold{\code{@multiLocalZscores}}
\item \bold{\code{@matrix}}

}
}
\description{
Perform a multiple permutation test and local z-score calculation between a region set and
list of regions set (or GrangedList)
}
\details{
the permutation test core used in this function allows to change
\code{"randomize.function"} \code{\link{randomizeRegions}},
\code{\link{circularRandomizeRegions}}, \code{\link{resampleRegions}} or a
custom function), but use only an \code{"evaluation.function"}
\code{\link{numOverlaps}}
}
\examples{

\dontrun{

data(cw_Alien)

mlz_Alien_regB <- multiLocalZscore(A = AlienRSList$regB, Blist = AlienRSList)

summary(mlz_Alien_regB)

}

}
\seealso{
\code{\link{localZScore}}
}
