% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotLocalZScoreMatrix.R
\name{plotLocalZScoreMatrix}
\alias{plotLocalZScoreMatrix}
\title{Plot Local Z-Score Matrix}
\usage{
plotLocalZScoreMatrix <- function(mLz, lineColor = NA, interpolate = FALSE, colMatrix = "default", matrix.type = "association" , maxVal = 2, main = "",  revert = FALSE)
}
\arguments{
\item{mLz}{an object of class multiLocaZScore or a matrix}

\item{lineColor}{logic if TRUE a grid matrix will be draw (default: FALSE)}

\item{interpolate}{logic the image will be interpolate using the function (\code{\link{geom_raster}}}

\item{colMatrix}{character or vector of colors, if "default" will be used a default selection see..}

\item{matrix.type}{character ("association" or "correlation") is the kind of matrix that will be plotted (default = "association")}

\item{maxVal}{numeric, maximum abs(value) reached by the plot. (default = 2)}

\item{main}{character, plot title}
}
\description{
Plot Local Z-Score Matrix of associations/correlations stored in a multiLocaZScore object.
}
