% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regioneReloaded_theme.R
\name{plotPal}
\alias{plotPal}
\title{plotPal}
\usage{
plotPal(colPal = NULL)
}
\arguments{
\item{colPal}{character, colors to use as palette for the plot. If NULL, default colors will be used. (default = NULL)}
}
\value{
A palette function
}
\description{
Internal function to create a palette generating function for plots
}
