% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genoMatriXeR_methods.R
\name{print,genoMatriXeR-method}
\alias{print,genoMatriXeR-method}
\title{Prints gMXR object}
\usage{
\S4method{print}{genoMatriXeR}(x)
}
\arguments{
\item{x}{A gMXR object}
}
\value{
a printing for genoMatriXeR objects
}
\description{
Prints gMXR object
}
