% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resampleGenome.R
\name{resampleGenome}
\alias{resampleGenome}
\title{resampleGenome}
\usage{
resampleGenome(A, simple = FALSE, per.chromosome = FALSE, genome="hg19", ...)
}
\arguments{
\item{A}{an object of class GenomigRanges}

\item{simple}{logical, if is TRUE the randomization process will not take care about the singular whit for each regions in A. (defalut = FALSE)}

\item{per.chromosome}{logical, if TRUE the randomization will be perform by chromosome. (default  = TRUE)}

\item{genome}{character or GenomicRanges, genome using for the randomization}

\item{...}{further arguments to be passed to other methods.}
}
\value{
a permTest object
}
\description{
function for use resample starting from a binning of genome
}
\seealso{
regioneR::permTest
}
