% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/similarRegionSet.R
\name{similarRegionSet}
\alias{similarRegionSet}
\title{similar RegionSet}
\usage{
similarRegionSet<-function(GR, name,genome,vectorPerc)
}
\arguments{
\item{GR}{an object of class gMXR or a matrix}

\item{name}{character, name of the first element of gMXR object to test.}

\item{genome}{character, name of the second element of gMXR object to test.}

\item{vectorPerc}{character, label for x axes. (default = NA)}
}
\description{
create a similar region set.
}
