% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subList.R
\name{subList}
\alias{subList}
\title{Sub List from region set list}
\usage{
subList(Alist, min_sampling, fraction)
}
\arguments{
\item{Alist, }{GRangesList or list of Region Set of any accepted formats by  \code{\link{regioneR}} package}

\item{min_sampling}{numeric, minimun number of regions for each element of the list, if the number of regions is less than min_samplig as subsempilg will be take all the regionset}

\item{fraction}{numeric, percentage of regions selected for the subsampling}
}
\description{
crerate a samples sublist starting from a list of Regin  Set
}
\keyword{function}
\keyword{internal}
