% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chooseHclustMet.R
\name{chooseHclustMet}
\alias{chooseHclustMet}
\title{Choose HClust Method}
\usage{
chooseHclustMet(GM, scale = FALSE, vecMet = NA,distHC = "euclidean")
}
\arguments{
\item{GM}{matrix,  numerical matrix from gMXR object.}

\item{scale}{logic, if TRUE, the clusterization will be perform using the scale version of the mnatrix. (default = FALSE)}

\item{vecMet}{vector, vector of method that will be tested from the function
if NULL the follow methods will be tested "complete","average","single","ward.D2","median","centroid" and "mcquitty. (default = NULL)}

\item{distHC}{character, algorithm use for calculate distance. (default = "euclidean")}
}
\description{
evaluate and choose the better method for clusterize a matrix using hclust function
}
\keyword{function}
\keyword{internal}
