% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleanCrosswiseMatrix.R
\name{cleanCrosswiseMatrix}
\alias{cleanCrosswiseMatrix}
\title{Clean Crosswise Matrix}
\usage{
cleanCrosswiseMatrix(GM, GM_pv, pvcut,scale)
}
\arguments{
\item{GM}{matrix,  numerical matrix from gMXR object.}

\item{GM_pv}{matrix, numerical matrix that represent p-value from gMXR object.}

\item{pvcut}{maximum pvalue accepted, (all the z-score vales with pvalue higher than "pvcut" will be transform in 0)}

\item{logic, }{if the matrix GM will be scaled using the function \code{link{scale}}. (default = FALSE)}
}
\description{
clean and scale a matrix from a gMXR object
}
\keyword{internal}
