% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createUniverse.R
\name{createUniverse}
\alias{createUniverse}
\title{Create Universe}
\usage{
createUniverse(Alist)
}
\arguments{
\item{Alist}{list of regions set in format accepted from \code{\link{regioneR}}}
}
\value{
A list of Genomic Ranges objects
}
\description{
create the universe parameter for \code{\link{resampleRegions}} function using all unique region present in Alist
}
\examples{

data("cw_Alien")

universe <- createUniverse(AlienRSList)

}
\keyword{internal}
