% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crosswiseMatrix.R
\name{crosswiseMatrix}
\alias{crosswiseMatrix}
\title{Matrix from crosswisePermTest Object}
\usage{
crosswiseMatrix(mPT, zs.type='norm_zscore',...)
}
\arguments{
\item{zs.type}{Character, choose if create the matrix using every column from MultiOverlap slot of gMXR object default: 'norm_zscore'}

\item{mPt, }{object of class = "gMXR"}
}
\description{
create a matrix from a gMXR object
}
\keyword{function}
\keyword{internal}
