% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crosswisePermTest.R
\name{crosswisePermTest}
\alias{crosswisePermTest}
\title{Multiple Permutation test}
\usage{
crosswisePermTest(Alist, Blist, sampling=FALSE, fraction=0.15,
                         ranFUN = "randomizeRegions",evFUN = "numOverlaps",
                         universe = NULL, adj_pv_method = "BH", max_pv = 0.05,
                         verbose = FALSE, subEx = 0, genome = "hg19",
                         verbose = FALSE...)
}
\arguments{
\item{Alist}{GRangesList or list of Region Set of any accepted formats by  \code{\link{regioneR}} package
(\code{\link{GenomicRanges}}, \code{\link{data.frame}} etc...)}

\item{Blist}{GRangesList or list of Region Set of any accepted formats by  \code{\link{regioneR}} package
(\code{\link{GenomicRanges}}, \code{\link{data.frame}} etc...)}

\item{sampling}{Boolean, if is true the function will use only a sample of
each element of Alist to perform the test (default = FALSE)}

\item{fraction}{Logic, if sampling==TRUE is the fraction of the region sets
used to perform the test (default = 0.15)}

\item{min_sampling}{numeric, minimum number of regions accepted after the sampling, if the number of the sampled
regions is less than min_sampling will be used min_sampling value as number of regions}

\item{ranFUN}{(default = "randomizeRegions") choose the randomization strategy used for the test see  \code{\link{regioneR}}}

\item{evFUN}{(default = "numOverlaps) choose the evaluation strategy used for the test see  \code{\link{regioneR}}}

\item{universe}{(default = NULL) used only when \code{\link{resampleRegions}} function is selected}

\item{adj_pv_method}{Charachter, the method used for the calculation of the adjusted p-value,
to choose between the options of \code{\link{p.adjust}}. (default = "BH")}

\item{max_pv}{Numeric, the z-scores associate a p-values higher of this parameter will be transform in subEx. (default =0.05)}

\item{subEx}{Numeric, (default = 0) substitute this value to a z-score when the p-value is higher than max_pv}

\item{genome}{Charachter or GenomicRanges, (defalut = "hg19") genome used to compute the randomization}

\item{verbose}{Boolean, if verbose test}
}
\value{
A object of class \code{gMXR}
}
\description{
Perform a multiple permutation test between each elements of 2 list of Region
set the result will be store in a gMXR S4 class
}
\details{
...
}
\examples{
 ...
}
\seealso{
\code{\link{regioneR}}, \code{\link{permTest}}, \code{\link{overlapPermTest}}
}
