% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regioneReloaded_data.R
\docType{data}
\name{cw_Alien}
\alias{cw_Alien}
\title{cw_Alien}
\format{
An objects of class \link[=genoMatriXeR-class]{genoMatriXeR}; see \code{\link[=makeCrosswiseMatrix]{makeCrosswiseMatrix()}}.
}
\usage{
data(cw_Alien)
}
\description{
Alien Genome crosswise matrix using \link[regioneR:randomizeRegions]{regioneR::randomizeRegions} , \link[regioneR:circularRandomizeRegions]{regioneR::circularRandomizeRegions},
\link[regioneR:resampleRegions]{regioneR::resampleRegions}, \link[regioneR:resampleGenome]{regioneR::resampleGenome} functions as permutation strategies.
}
\keyword{datasets}
