% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cw_Alien_RaR-data.R
\docType{data}
\name{cw_Alien_RaR}
\alias{cw_Alien_RaR}
\title{Alien Genome crosswise matrix using \code{\link{randomizeRegions}} function as permutation strategy}
\format{
An object of class \code{"genoMatriXeR"}; see \code{\link{makeCrosswiseMatrix}}.
}
\usage{
data(cw_Alien_RaR)
}
\description{
CreossWise matrix calculated using 39 di fferent regionset vs itself
(description of Region set can be find here:"")
}
\keyword{datasets}
