% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regioneReloaded_data.R
\docType{data}
\name{cw_Alien_ReG}
\alias{cw_Alien_ReG}
\title{cw_Alien_ReG}
\format{
An objects of class \link[=genoMatriXeR-class]{genoMatriXeR}; see \code{\link[=makeCrosswiseMatrix]{makeCrosswiseMatrix()}}.
}
\usage{
data(cw_Alien)
}
\description{
Alien Genome crosswise matrix using \code{\link[=resampleGenome]{resampleGenome()}} function a permutation s
trategy.
}
\keyword{datasets}
