% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genoMatriXeR_classes.R
\docType{class}
\name{genoMatriXeR-class}
\alias{genoMatriXeR-class}
\alias{gMXR}
\title{genoMatriXeR Class}
\description{
An S4 class for "genoMatriXeR" object.
}
\section{Slots}{

\describe{
\item{\code{parameters}}{List of parameters used to create the object.}

\item{\code{multiOverlaps}}{Results of multiple pairwise permutation tests generated with \code{\link[=crosswisePermtest]{crosswisePermtest()}}.}

\item{\code{matrix}}{List of numerical matrices containing z-score, pvalues and correlation values generated with \code{\link[=makeCrosswiseMatrix]{makeCrosswiseMatrix()}}}
}}

