% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genoMatriXeR_classes.R
\docType{class}
\name{genoMatriXeR-class}
\alias{genoMatriXeR-class}
\alias{gMXR}
\title{An S4 class for "genoMatriXeR" object.}
\description{
An S4 class for "genoMatriXeR" object.
}
\section{Slots}{

\describe{
\item{\code{parameters}}{List of parameters used to create the object}

\item{\code{multiOverlaps}}{...}

\item{\code{matrix}}{...}
}}

