% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMatrix.R
\name{getMatrix}
\alias{getMatrix}
\title{Get Matrix}
\usage{
getMatrix(rR)
}
\arguments{
\item{rR}{genoMatriXeR or multiLocalZScore object}
}
\value{
a numerical matrix from a
}
\description{
Returns the matrix from an \link[=genoMatriXeR-class]{genoMatriXeR} or \link[=multiLocalZScore-class]{multiLocalZScore} object.
}
\examples{

data("cw_Alien")

cw_Alien_ReG <- makeCrosswiseMatrix(cw_Alien_ReG)
mtx <- getMatrix(cw_Alien_ReG)

mtx


data("cw_Alien")

cw_Alien_RaR <- makeCrosswiseMatrix(cw_Alien_RaR)
GM <- getMatrix(cw_Alien_RaR)

GM

}
\seealso{
\link[=genoMatriXeR-class]{genoMatriXeR}, \link[=multiLocalZScore-class]{multiLocalZScore}, \link{makeCrosswiseMatrix}, \link{makeLZMatrix}
}
