% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMultiEvaluation.R
\name{getMultiEvaluation}
\alias{getMultiEvaluation}
\title{getMultiEvaluation}
\usage{
getMultiEvaluation( rR, namesRS = NA)
}
\arguments{
\item{rR}{A genoMatriXeR or multiLocalZScore object.}

\item{namesRS}{a vector of names. (default = NA)}
}
\value{
If rR is a \link[=genoMatriXeR_class]{genoMatriXeR} object a list of data frames resuming the associations results.
If rR is a \link[=multiLocalZScore_class]{multiLocalZScore} object return al list of two elements "resumeTable" that is the resume of the associations analysis and "shifts",
a list of shifts computed from \code{\link[=multilocalZScore]{multilocalZScore()}} function for the element indicated in the nameRS vector.
}
\description{
get multiEvaluation Object from genoMatriXeR or multiLocalZScore class
}
\seealso{
\link[=genoMatriXeR-class]{genoMatriXeR}
}
