% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getParameters.R
\name{getParameters}
\alias{getParameters}
\title{getParameters}
\usage{
getParameters(rR)
}
\arguments{
\item{rR}{A genoMatriXeR or multiLocalZScore class object.}
}
\value{
a dataframe with parameters and values.
}
\description{
get Parametes from genoMatriXeR or multiLocalZScore class
}
\examples{

data("cw_Alien")

prm <- getParameters(cw_Alien_ReG)

prm


}
\seealso{
\link[=genoMatriXeR-class]{genoMatriXeR}, \link[=multiLocalZScore-class]{multiLocalZScore}
}
