% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getParameters.R
\name{getParameters}
\alias{getParameters}
\title{getParameters}
\usage{
getParameters(rR, show_err = FALSE)
}
\arguments{
\item{rR}{A genoMatriXeR or multiLocalZScore class object.}

\item{show_err}{logical, if TRUE the function returns a list with two dataframes:
one containing the parameter values and one with any error messages that have been
generated during the permutation test iterations when running \link{crosswisePermTest}.}
}
\value{
A dataframe with parameters and values, or a list with two dataframes
with parameters and errors information.
}
\description{
Get parameters from a genoMatriXeR or multiLocalZScore class object.
}
\examples{

data("cw_Alien")

prm <- getParameters(cw_Alien_ReG)

prm


}
\seealso{
\link[=genoMatriXeR-class]{genoMatriXeR}, \link[=multiLocalZScore-class]{multiLocalZScore}
}
