% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeCrosswiseMatrix.R
\name{makeCrosswiseMatrix}
\alias{makeCrosswiseMatrix}
\title{Make Crosswise Matrix}
\usage{
makeCrosswiseMatrix( mPt, clusterize = TRUE, hc.method = NA,dist.method = "euclidean",transform = FALSE,
                           scale = FALSE, zs.type = 'norm_zscore', symm_matrix = TRUE, selectVec = NULL,
                           maxNZS = 100, pvcut = 0.05, ...)
}
\arguments{
\item{clusterize}{logic, if TRUE the matrix will be clusterize using a method selected with the variable \code{hc.method} (default = TRUE)}

\item{hc.method}{character, select the \code{hclust} method to use for clusterize the matrix,
if hc.method == NA, the methods used for clusterize the matrix will be selected using the function \code{\link{chooseHclustMet}} (default = "NA")}

\item{dist.method}{character, metric used to calculate the distance matrix see \code{hclust} (default = "euclidean")}

\item{transform}{logic, if TRUE the matrix will be transform using the function \code{("t")} (default = FALSE)}

\item{scale}{logic, if TRUE the matrix will be scaled (default = FALSE)}

\item{zs.type}{character, choose if create the matrix using every column from MultiOverlap slot of gMXR object default: 'norm_zscore'}

\item{symm_matrix}{logic, if TRUE the matrix will be treated as symmetrical (same clusterization for row and column). (default = TRUE)}

\item{selectVec}{vector, teh matrix will be reduced using only the row/column content. (default = NULL)}

\item{pvcut}{maximum adj.pvalue accepted, all the associations with a adj.pvalue (defined in \code{\link{crosswisePermTest}}) higher than pvcut were transformed in 0. (default = 0.05)}

\item{...}{urther arguments to be passed to other methods.}

\item{mPt}{an object of class gMXR or a matrix}
}
\description{
Accept as input a gMXR object an create in the slot Matrix all the fields.
}
