% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeLZMatrix.R
\name{makeLZMatrix}
\alias{makeLZMatrix}
\title{Make Local Z-Score Matrix}
\usage{
makeLZMatrix( mlZA, normalize = TRUE, clusterize = TRUE, hc.method = NULL, dist.method = "euclidean",
                           scale = FALSE, ...)
}
\arguments{
\item{mlZA}{an object of class \code{\link{multiLocalZScore }} or a matrix}

\item{normalize}{logic, if TRUE the matrix will be normalize. (default = FALSE)}

\item{clusterize}{logic, if TRUE the matrix will be clusterize using a method selected with the variable \code{hc.method} (default = TRUE)}

\item{centralize}{numeric, number of "steps" around the center of the local association in which will be apply the clusterization algorithm, if NA all the values in the matrix will be utilize. (default = NA)}

\item{hc.method}{character/vector, select the \code{hclust} method to use for clusterize the matrix,
if hc.method == NULL, the methods used for clusterize the matrix will be selected using the function \code{\link{chooseHclustMet}} (default = "NA")}

\item{dist.method}{character, metric used to calculate the distance matrix see \code{hclust} (default = "euclidean")}

\item{scale}{logic, if TRUE the matrix will be scaled (default = FALSE)}
}
\description{
create a local z-score matrix from an \code{\link{multiLocalZScore }} object and save it in the matrix slot
}
