% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regioneReloaded_data.R
\docType{data}
\name{mLZ_regA_ReG}
\alias{mLZ_regA_ReG}
\title{mLZ_regA_ReG}
\format{
An objects of class \link[=multiLocalZScore-class]{multiLocalZScore}; see \code{\link[=makeLZMatrix]{makeLZMatrix()}}.
}
\usage{
data(cw_Alien)
}
\description{
Alien Genome multiLocalZScore calculated for regA regionset from AlienRSList_narrow using \code{\link[regioneR:resampleGenome]{regioneR::resampleGenome()}} function as permutation s
trategy.
}
\keyword{datasets}
