% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genoMatriXeR_classes.R
\docType{class}
\name{multiLocalZScore-class}
\alias{multiLocalZScore-class}
\alias{mLZS}
\title{class multiLocalZScore}
\description{
multiLocalZScore for testing
}
\details{
@slot Name a character.
 @export multiLocalZScore
}
