% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mLZS_class.R
\docType{class}
\name{multiLocalZScore-class}
\alias{multiLocalZScore-class}
\alias{mLZS}
\title{multiLocalZScore Class}
\description{
An S4 class for "multiLocalZScore" object.
}
\section{Slots}{

\describe{
\item{\code{parameters}}{List of parameters used to create the object}

\item{\code{multiLocalZscores}}{...}

\item{\code{matrix}}{...}
}}

