% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiLocalZscore.R
\name{multiLocalZscore}
\alias{multiLocalZscore}
\title{Multiple Local Z-Score test}
\usage{
listOvelapPermtest( Alist, Blist, sampling=FALSE, fraction=0.15,
ranFun="randomizeRegions", ...)
}
\arguments{
\item{A}{Region Set of any accepted formats by  \code{\link{regioneR}} package
(\code{\link{GenomicRanges}}, \code{\link{data.frame}} etc...)}

\item{Blist}{list of Region Set of any accepted formats by \code{\link{regioneR}} package
(\code{\link{GenomicRanges}}, \code{\link{data.frame}} etc...)}

\item{sampling}{Boolean, if is true the function will use only a sample of
each element of Alist to perform the test}

\item{fraction}{Numeric, if sampling==TRUE is the fraction of the region sets
used to perform the test}

\item{universe}{Region Set of any accepted formats by  \code{\link{regioneR}}, using only when resamplinRegions function is
selected (default = NULL)}

\item{adj_pv_method}{Charachter, the method used for the calculation of the adjusted p-value,
to choose between the options of \code{\link{p.adjust}}. (default = "BH")}

\item{max_pv}{Numeric, the z-scores associate a p-values higher of this parameter will be transform in 0. (default =0.05)}

\item{...}{further arguments to be passed to other methods.}

\item{ranFun}{Function, choose the randomization strategy used for the test (default)
for details see \code{\link{regioneR}}}

\item{verbose}{Boolean, if verbose}
}
\value{
...
}
\description{
Perform a multiple permutation test and local z-score calculation between a region set and
list of regions set (or GrangedList)
}
\details{
the permutation test core used in this function permit to change
\code{"randomize.function"} \code{\link{randomizeRegions}},
\code{\link{circularRandomizeRegions}}, \code{\link{resampleRegions}} or a
custom function), but use only an \code{"evaluation.function"}
\code{\link{numOverlaps}}
}
\examples{
 ...

}
\seealso{
\code{\link{localZscore}}
}
