% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiPermTest.R
\name{multiPermTest}
\alias{multiPermTest}
\title{multiPermTest}
\usage{
multiPermTest (A, Blist, ranFUN, evFUN, universe, genome, rFUN, ntimes, adj_pv_method, ...)
}
\arguments{
\item{A}{Genomic Ranges or any accepted formats by  \href{https://bioconductor.org/packages/release/bioc/html/regioneR.html}{regioneR} package
(\code{\link{GenomicRanges}}, \code{\link{data.frame}} etc...)}

\item{ranFUN}{character, the randomization strategy used for the test, see \href{https://bioconductor.org/packages/release/bioc/html/regioneR.html}{regioneR}. (default = "randomizeRegions")}

\item{evFUN}{character, the evaluation strategy used for the test, see  \href{https://bioconductor.org/packages/release/bioc/html/regioneR.html}{regioneR}. (default = "numOverlaps)}

\item{universe}{region set to use as universe, used only when \code{\link[regioneR:resampleRegions]{regioneR::resampleRegions()}} function is selected. (default = NULL)}

\item{genome}{character or \linkS4class{GRanges}, genome used to compute the randomization. (default = "hg19")}

\item{ntimes}{numeric, number of permutations used in the test. (default = 100)}

\item{adj_pv_method}{character, the method used for the calculation of the adjusted p-value, to choose between the options of \code{\link[=p.adjust]{p.adjust()}}. (default = "BH")}

\item{...}{further arguments to be passed to other methods.}
}
\value{
a data frame object computed starting from results of \code{\link[regioneR:permTest]{regioneR::permTest()}} function
}
\description{
multiPermTest
}
\details{
Perform a multiple permutation test
}
\seealso{
\code{\link[regioneR:permTest]{regioneR::permTest()}}
}
\keyword{function}
\keyword{internal}
