% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiPermTest.R
\name{multiPermTest}
\alias{multiPermTest}
\title{Multiple Permutation Test}
\usage{
multiPermTest A, Blist, ranFUN, evFUN, universe, genome, rFUN, verbose = FALSE, ntimes, adj_pv_method, ...)
}
\description{
Perform a multiple permutation test
}
\keyword{function}
\keyword{internal}
