% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotCrosswiseMatrix.R
\name{plotCrosswiseMatrix}
\alias{plotCrosswiseMatrix}
\title{Plot Crosswise Matrix}
\usage{
plotCrosswiseMatrix(mPt,  lineColor = NA, interpolate = FALSE, colMatrix = "default", matrix.type = "crosswise", cor = "row", main = "")
}
\arguments{
\item{mPt}{an object of class gMXR or a matrix}

\item{lineColor}{logic if TRUE a grid matrix will be draw (default: FALSE)}

\item{interpolate}{logic the image will be interpolate using the function (\code{\link{geom_raster}}}

\item{colMatrix}{character or vector of colors, if "default" will be used a default selection see..}

\item{matrix.type}{character ("crosswise" or "correlation") is the kind of matrix that will be plotted.}

\item{cor}{character ("row" or "col"). If matrix variable is "correlation", defined if the function \code{"cor"} will be executed on rows or columns of the matrix}

\item{maxVal}{numeric, maximun abs(value) reached by the plot. (default = 2)}

\item{main}{charachter, graph title}
}
\description{
Plot matrix of associations/correlations stored in a gMXR object.
}
