% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotCrosswiseMatrix.R
\name{plotCrosswiseMatrix}
\alias{plotCrosswiseMatrix}
\title{Plot Crosswise Matrix}
\usage{
plotCrosswiseMatrix(mPT, lineColor = NA, interpolate = FALSE, colMatrix =
"default", matrix_type = "crosswise", cor = "row",
maxVal = NA, main = "", ord_mat = NULL)
}
\arguments{
\item{mPT}{an object of class gMXR or a matrix}

\item{lineColor}{logic if TRUE a grid matrix will be draw (default: FALSE)}

\item{interpolate}{logic the image will be interpolate using the function (\code{\link{geom_raster}}}

\item{colMatrix}{character or vector of colors, if "default" will be used a default selection see..}

\item{matrix_type}{character ("crosswise" or "correlation") is the kind of matrix that will be plotted.}

\item{cor}{character ("row" or "col"). If matrix variable is "correlation", defined if the function \code{"cor"} will be executed on rows or columns of the matrix}

\item{maxVal}{numeric, maximun abs(value) reached by the plot. (default = 2)}

\item{main}{charachter, graph title}

\item{ord_mat}{numeric list, list with two vectors X and Y as element, that
represent the order of rows and column of the matrix (default = NULL)}
}
\value{
A plot is created on the current graphics device.
}
\description{
Plot matrix of associations/correlations stored in a genoMatriXeR object.
}
\examples{

data("cw_Alien")

cw_Alien_ReG <- makeCrosswiseMatrix( cw_Alien_ReG)

plotCrosswiseMatrix(cw_Alien_ReG, matrix_type = "crosswise")

plotCrosswiseMatrix(cw_Alien_ReG, matrix_type = "correlation")


}
\seealso{
\code{\link{crosswisePermTest}} \code{\link{makeCrosswiseMatrix}}
}
