% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotCrosswiseMatrix.R
\name{plotCrosswiseMatrix}
\alias{plotCrosswiseMatrix}
\title{plotCrosswiseMatrix}
\usage{
plotCrosswiseMatrix(mPT, lineColor = NA, interpolate = FALSE, colMatrix =
"default", matrix_type = "crosswise", cor = "row",
maxVal = NA, main = "", ord_mat = NULL)
}
\arguments{
\item{mPT}{an object of class \link[=genomeMatriXeR-class]{genoMatriXeR} or a numerical matrix.}

\item{lineColor}{logical, color for the line grid delineating the tiles of the matrix plot. If NA, no line will be drawn. (default = NA)}

\item{interpolate}{logical, if TRUE the image will be interpolated using the function \code{\link[=geom_raster]{geom_raster()}}. (default = FALSE)}

\item{colMatrix}{character or vector of colors, if "default" will be used a default selection see..}

\item{matrix_type}{character, type of matrix to be plotted, either "crosswise" or "correlation". (default = "crosswise)}

\item{cor}{character, if \code{matrix_type} is "correlation", choose if the function \code{\link[=cor]{cor()}} will be executed on each "row" or "column" of the matrix. (default = "row")}

\item{maxVal}{numeric, maximum absolute value displayed by the plot. If NA, the 0.95 quantile of all absolute values is used. (default = NA)}

\item{main}{character, title of the plot. (default = "")}

\item{ord_mat}{numeric, list with two numeric vectors that represent the ordering of rows and column of the matrix to be used in the plot.
If NULL, the order of the matrix is preserved as is. (default = NULL)}
}
\value{
Returns a ggplot object.
}
\description{
Plot matrix of associations/correlations stored in a \link[=genomeMatriXeR-class]{genoMatriXeR} object.
}
\examples{

data("cw_Alien")

cw_Alien_ReG <- makeCrosswiseMatrix( cw_Alien_ReG)

plotCrosswiseMatrix(cw_Alien_ReG, matrix_type = "crosswise")

plotCrosswiseMatrix(cw_Alien_ReG, matrix_type = "correlation")


}
\seealso{
\code{\link{crosswisePermTest}} \code{\link{makeCrosswiseMatrix}}
}
