% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotLocalZScoreMatrix.R
\name{plotLocalZScoreMatrix}
\alias{plotLocalZScoreMatrix}
\title{Plot Local Z-Score Matrix}
\usage{
plotLocalZScoreMatrix (mLZ, lineColor = NA, colMatrix = "default",
matrix.type = "association", maxVal = "max", main = "", labSize = 6,
revert = FALSE, highlight = NULL, highlight_size = 2.5, highlight_max = FALSE,
smoothing = FALSE, ...)
}
\arguments{
\item{mLZ}{an object of class \linkS4class{multiLocalZScore} or a matrix}

\item{lineColor}{logic if TRUE a grid matrix will be draw (default: FALSE)}

\item{colMatrix}{character or vector of colors, if "default" will be used a default selection see..}

\item{matrix.type}{character ("association" or "correlation") is the kind of matrix that will be plotted (default = "association")}

\item{maxVal}{numeric, maximum abs(value) reached by the plot. (default = 2)}

\item{main}{character, plot title}

\item{labSize}{numeric, size for the plot lab}

\item{revert}{logic, revert the order of the plotted elements}

\item{highlight}{character vector indicating the region set names to highlight by adding labels pointing to the 0 shift position (default = NULL)}

\item{highlight_size}{numeric, size of the highlight labels.}

\item{highlight_max}{logical, if TRUE the highlight labels are placed at the maximum local z-score value instead of the 0 shift position. (default = FALSE)}

\item{smoothing}{logical, if TRUE the \link[stats:smooth.spline]{stats::smooth.spline} function will be applied to the local z-score profile. (default = FALSE)}

\item{...}{further arguments to be passed to other methods.}
}
\value{
Returns a ggplot object.
}
\description{
Plot Local Z-Score Matrix of associations/correlations stored in a \linkS4class{multiLocalZScore} object.
}
\examples{

data("cw_Alien")



}
\seealso{
\link{multiLocalZscore} \link{makeLZMatrix} \linkS4class{multiLocalZScore}
}
