% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotLocalZScoreMatrix.R
\name{plotLocalZScoreMatrix}
\alias{plotLocalZScoreMatrix}
\title{Plot Local Z-Score Matrix}
\usage{
plotLocalZScoreMatrix (mLZ, lineColor = NA, colMatrix = "default",
matrix.type = "association" , maxVal = 2, main = "", labSize= 6, revert = FALSE,
highlight = NULL, highlight_size = 2.5)
}
\arguments{
\item{mLZ}{an object of class multiLocaZScore or a matrix}

\item{lineColor}{logic if TRUE a grid matrix will be draw (default: FALSE)}

\item{colMatrix}{character or vector of colors, if "default" will be used a default selection see..}

\item{matrix.type}{character ("association" or "correlation") is the kind of matrix that will be plotted (default = "association")}

\item{maxVal}{numeric, maximum abs(value) reached by the plot. (default = 2)}

\item{main}{character, plot title}

\item{labSize}{numeric, size for the plot lab}

\item{revert}{logic, revert the order of the plotted elements}

\item{highlight}{character vector indicating the regionset names to highlight by adding labels pointing to the 0 position (default = NULL)}

\item{highlight_size}{numeric, size of the highlight labels}
}
\value{
A plot is created on the current graphics device.
}
\description{
Plot Local Z-Score Matrix of associations/correlations stored in a multiLocaZScore object.
}
\examples{

data("cw_Alien")

plotLocalZScoreMatrix(mLZ_Alien_ReG)


}
\seealso{
\code{\link{multiLocalZScore}} \code{\link{makeLZMatrix}}
}
