% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotSingleLZ.R
\name{plotSingleLZ}
\alias{plotSingleLZ}
\title{plotSingleLZ}
\usage{
plotSingleLZ(mLZ, RS, xlab = "", normZS = TRUE, ylim = NULL,
main = NA, colPal = NULL, labValues = TRUE, labSize = 2.5)
}
\arguments{
\item{mLZ}{an object of class multiLocalZscore.}

\item{RS}{character, vector of region set names for which to plot the local Z-score results.}

\item{xlab}{character, label for the x axis. (default = NA)}

\item{normZS}{logical, indicates whether the normalized Z-score values should be plotted. If FALSE, the raw Z-score is used. (default = TRUE)}

\item{ylim}{numeric, vector with minimum and maximum Y values of the plot. If NULL, the plot limits are set by default so all data points can be plotted. (default = NULL)}

\item{main}{character, title for the plot. (default = NA)}

\item{colPal}{character vector of custom colors to use as palette source for the plot. If NULL, predetermined colors from \code{\link{RColorBrewer}}) Set2 palette are used.}

\item{labValues}{logical, if TRUE each local Z-score profile is labelled at position 0 with the name of the region set and its Z-score value at the central position. (default = TRUE)}

\item{labSize}{numerical, size of the labels from labValues in the plot. (default = 2.5)}
}
\value{
A plot is created on the current graphics device.
}
\description{
Plot the result of specific local Z-Score tests from an mLZ object in the
form of line plot profiles.
}
\details{
This function generates a line plot with the local Z-score profiles of
selected region sets from a multiLocalZscore object. This type of plot
complements the local Z-score matrix (generated by
\code{\link[RegioneReload]{plotLocalZScoreMatrix}}), since it allows to
visualize in detail the local Z-score profile of just the region sets of
interest. This plot is well suited for a single or a few region sets, but
will get busy if attempting to plot many different profiles. For the latter,
the full matrix generated by \code{\link[RegioneReload]{plotLocalZScoreMatrix}})
is usually a better visualization option.
}
\examples{

data("cw_Alien")

# plotSingleLZ(mlz_Alien_ReG, RS = c("regAB")


}
\seealso{
\code{\link{multiLocalZScore}} \code{\link{makeLZMatrix}}
}
