% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotSingleLZ.R
\name{plotSingleLZ}
\alias{plotSingleLZ}
\title{Plot Single Local Z-Score Test}
\usage{
plotSingleLZ<-function(mLZ, RS2,  colvec = NA, xlab = " N. of overlaps", limH = NA, main = NA)
}
\arguments{
\item{mLZ}{an object of class multilocalZscore}

\item{RS}{character, name of regionSet to test.}

\item{xlab}{character, label for x axes. (default = NA)}

\item{normZS}{logical, indicates whether the normalized Z-score values should be plotted. If FALSE, the raw
Z-score is used (default = TRUE).}

\item{main}{character, title for the plot. (default = NA)}

\item{colpal}{character vector of custom colors to use as palette source for the plot. If NULL, predetermined
colors are used.}

\item{labValues}{logical, if TRUE retrieve the value of z-score and adjusted pvaue for each tested associations}

\item{label_size}{numerical, size for the label in the plot}
}
\value{
A plot is created on the current graphics device.
}
\description{
Plot the result of a single local Z-Score test from an mLZ object.
}
\examples{

data("cw_Alien")

plotSingleLZ(mlz_Alien_Reg, RS = "regAB")


}
\seealso{
\code{\link{multiLocalZScore}} \code{\link{makeLZMatrix}}
}
