% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotSinglePT.R
\name{plotSinglePT}
\alias{plotSinglePT}
\title{Plot Single Permutation Test}
\usage{
plotSinglePT<-function(mPT, RS1, RS2, xlab = NA, main = NA)
}
\arguments{
\item{RS1}{character, name of the first element of genoMatriXeR object to test.}

\item{RS2}{character, name of the second element of genoMatriXeR object to test.}

\item{xlab}{character, label for x axis (default = NA)}

\item{mPt}{an object of class genoMatriXeR}
}
\value{
A plot is created on the current graphics device.
}
\description{
Plot the result of a single pairwise permutation test from a genoMatriXeR object.
}
\examples{

data("cw_Alien")
plotSinglePT(cw_Alien_ReG, RS1 = "regA", RS2 = "regA05")
plotSinglePT(cw_Alien_ReG, RS1 = "regA", RS2 = "regC")

}
\seealso{
\code{\link{crosswisePermTest}} \code{\link{makeCrosswiseMatrix}}
}
