% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotSinglePT.R
\name{plotSinglePT}
\alias{plotSinglePT}
\title{Plot Single Permutation Test}
\usage{
plotSinglePT<-function(mPT,RS1,RS2,xlab =" N. of overlaps",main=NA,add_theme = FALSE,colvec = NULL)
}
\arguments{
\item{RS1}{character, name of the first element of gMXR object to test.}

\item{RS2}{character, name of the second element of gMXR object to test.}

\item{xlab}{character, label for x axes. (default = NA)}

\item{main}{character, title for the plot. (default = NA)}

\item{add_theme}{logic, if TRUE will add a default \code{\link{ggplot2}} theme. (default = FALSE)}

\item{colvec}{vector, vector of colors used in the plot, if NULL it will chooose a default vector. (default = NULL)}

\item{mPt}{an object of class gMXR or a matrix}
}
\description{
Plot the result of a single permutation test from an gMXR object.
}
