% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regioneReloaded_methods.R
\name{print,genoMatriXeR-method}
\alias{print,genoMatriXeR-method}
\title{Print method for genoMatriXeR class}
\usage{
\S4method{print}{genoMatriXeR}(x)
}
\arguments{
\item{x}{A genoMatriXeR object.}
}
\value{
A printed output for genoMatriXeR objects.
}
\description{
Print method for genoMatriXeR class
}
\seealso{
\link[=genoMatriXeR-class]{genoMatriXeR}
}
\keyword{internal}
