% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randomizeRegionsPerc.R
\name{randomizeRegionsPerc}
\alias{randomizeRegionsPerc}
\title{randomize Regions Perc}
\usage{
randomizeRegionsPerc(GR, genome, frac)
}
\arguments{
\item{GR}{description}

\item{genome}{description}

\item{frac}{description}
}
\description{
randomize regions at percentage
}
