% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randomizeRegionsPerc.R
\name{randomizeRegionsPerc}
\alias{randomizeRegionsPerc}
\title{randomize Regions Perc}
\usage{
randomizeRegionsPerc(GR, genome = "hg19", frac = 0.2, ...)
}
\arguments{
\item{GR}{description}

\item{genome}{description}

\item{frac}{description}

\item{...}{further arguments to be passed to other methods.}
}
\description{
randomize regions at percentage
}
\examples{

data("cw_Alien")

nreg <- 100

regA <-
  createRandomRegions(
  nregions = nreg,
  length.mean = 100,
  length.sd = 10,
  non.overlapping = TRUE,
  genome = AlienGenome
 )

 regA_02 <- randomizeRegionsPerc(GR = regA, genome = AlienGenome, frac = 0.2)

}
\seealso{
similarRegionSet
}
