% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resampleGenome.R
\name{resampleGenome}
\alias{resampleGenome}
\title{resampleGenome}
\usage{
resampleGenome(A, simple = FALSE, per.chromosome = FALSE, genome="hg19", ...)
}
\arguments{
\item{A}{an object of class GenomigRanges}

\item{simple}{logical, if is TRUE the randomization process will not take care about the singular whit for each regions in A. (defalut = FALSE)}

\item{per.chromosome}{logical, if TRUE the randomization will be perform by chromosome. (default  = TRUE)}

\item{genome}{character or GenomicRanges, genome using for the randomization}

\item{...}{further arguments to be passed to other methods.}
}
\value{
a \code{\link{GenomicRanges}} object. A sample from the \code{universe} with the same length as A.
}
\description{
function for use resample starting from a binning of genome
}
\examples{

A <- data.frame(chr=1, start=c(2,12,28,35), end=c(5,25,33,43))

resampleGenome(A,per.chromosome=TRUE)


}
\seealso{
\code{\link{toDataframe}}, \code{\link{toGRanges}}, \code{\link{randomizeRegions}}, \code{\link{createRandomRegions}}
}
