% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/similarRegionSet.R
\name{similarRegionSet}
\alias{similarRegionSet}
\title{similar RegionSets}
\usage{
similarRegionSet<-function(GR, name,genome,vectorPerc)
}
\arguments{
\item{GR}{an object of class GenomigRanges}

\item{name}{character, name of the first element of gMXR object to test.}

\item{genome}{character, name of the second element of gMXR object to test.}

\item{vectorPerc}{character, label for x axes. (default = NA)}
}
\value{
A list of class GenomicRanges
}
\description{
Create a list of Genomic Ranges objects similar to a reference Region Set
}
\examples{

data("cw_Alien")

A<-createRandomRegions(nregions = 20, length.mean = 1000, length.sd = 100, genome = AlienGenome)
similAList <- similarRegionSet(GR = A, genome = AlienGenome,vectorPerc = seq(0.1,0.9,0.2), name = "test")
summary (similAlist)

}
\seealso{
\code{\link{GenomiCRanges}}
}
