% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subList.R
\name{subList}
\alias{subList}
\title{Sub List from region set list}
\usage{
subList(Alist, min_sampling, fraction)
}
\arguments{
\item{Alist, }{GRangesList or list of Region Set of any accepted formats by  \href{https://bioconductor.org/packages/release/bioc/html/regioneR.html}{regioneR} package}

\item{min_sampling}{numeric, minimum number of regions for each element of the list,
if the number of regions is less than min_samplnig as subsempilng will be take the entire regionset}

\item{fraction}{numeric, percentage of regions selected for the subsampling}
}
\value{
a sublist of GenomicRanges from an original GenomicRanges list
}
\description{
create a samples sublist starting from a list of Region Set
}
